Divide commit function
This will divide to commit function.
