+++ scripts/github/constants/actions.js
@@ -2,4 +2,5 @@ export const ACTIONS = {
   OPEN_OAUTH_PAGE: 'openOauthPage',
   REQUEST_AND_SAVE_ACCESS_TOKEN: 'requestAndSaveAccessToken',
   GET_REPOSITORIES: 'getRepositories',
+  COMMIT: 'commit',
 };
