+++ scripts/github/services/repoService.js
@@ -21,4 +21,43 @@ export const repoService = {
 
     return await response.json();
   },
+
+  async commit(payload) {
+    const accessToken = await Util.getChromeStorage('githubAccessToken');
+    if (Util.isEmpty(accessToken)) {
+      throw new Error(Github.ERROR[Github.INVALID_ACCESS_TOKEN]);
+    }
+
+    const githubID = await Util.getChromeStorage('githubID');
+    if (Util.isEmpty(githubID)) {
+      throw new Error(Github.ERROR[Github.INVALID_GITHUB_ID]);
+    }
+
+    const uploadedRepository = await Util.getChromeStorage(
+      'githubUploadedRepository',
+    );
+    if (Util.isEmpty(uploadedRepository)) {
+      throw new Error(Github.ERROR[Github.INVALID_UPLOADED_REPOSITORY]);
+    }
+
+    const { extension, problemId, sha, sourceCode, type, title } = payload;
+    const path = `${type}/${problemId}.${extension}`;
+    const url = `${GITHUB_CONFIG.API_BASE_URL}/repos/${githubID}/${uploadedRepository}/contents/${path}`;
+    const headers = {
+      accept: 'application/vnd.github+json',
+      Authorization: `Bearer ${accessToken}`,
+    };
+
+    const body = JSON.stringify({
+      content: Util.encodeBase64Unicode(sourceCode),
+      message: title,
+      sha,
+    });
+
+    const response = await Util.request(url, 'PUT', headers, body);
+    const text = await response.json();
+    const message = text.commit.html_url ?? text.message;
+
+    return message;
+  },
 };
